requirements.txt
llama-index
llama-parse
llama-index-embeddings-openai
python-dotenv